Extract `FeedImageDataLoader` test helper into shared protocol extension to remove duplication
